Move formatPercent down into metabase/utils/formatting
Severs the visualizations -> static-viz formatPercent imports by moving the
helper from static-viz/lib/numbers into the formatting primitives module,
where the signature test (plain value in, string out) says it belongs.
