+++ enterprise/frontend/src/metabase-enterprise/transforms-inspector/pages/TransformInspectPage/components/LensSections/GenericSummarySections/components/FieldInfoSection/utils.tsx
@@ -2,7 +2,7 @@ import type { CellContext } from "@tanstack/react-table";
 import { c, t } from "ttag";
 
 import { getFormattedTime } from "metabase/common/components/DateTime";
-import { formatNumber, formatPercent } from "metabase/static-viz/lib/numbers";
+import { formatNumber } from "metabase/static-viz/lib/numbers";
 import { Ellipsified } from "metabase/ui";
 import {
   EntityNameCell,
@@ -10,6 +10,7 @@ import {
   Text,
   type TreeTableColumnDef,
 } from "metabase/ui";
+import { formatPercent } from "metabase/utils/formatting";
 import {
   isDate,
   isDateWithoutTime,
