+++ frontend/src/metabase/static-viz/components/FunnelChart/utils/funnel.ts
@@ -1,8 +1,9 @@
 import type { PolygonProps } from "@visx/shape/lib/shapes/Polygon";
 
 import { CHAR_SIZES_FONT_WEIGHT } from "metabase/static-viz/constants/char-sizes";
-import { formatNumber, formatPercent } from "metabase/static-viz/lib/numbers";
+import { formatNumber } from "metabase/static-viz/lib/numbers";
 import { measureTextWidth } from "metabase/static-viz/lib/text";
+import { formatPercent } from "metabase/utils/formatting";
 import type { TextWidthMeasurer } from "metabase/utils/measure-text";
 import { isNotNull } from "metabase/utils/types";
 import type {
