+++ frontend/src/metabase/static-viz/lib/numbers.ts
@@ -35,6 +35,3 @@ export const formatNumber = (number: number, options?: NumberFormatOptions) => {
 
   return `${prefix}${appFormatNumber(number, optionsWithDefault)}${suffix}`;
 };
-
-export const formatPercent = (percent: number) =>
-  `${(100 * percent).toFixed(Math.abs(percent) === 1 ? 0 : 2)} %`;
