+++ frontend/src/metabase/visualizations/components/ChartTooltip/TooltipRow/TooltipRow.tsx
@@ -1,7 +1,7 @@
 import cx from "classnames";
 import { t } from "ttag";
 
-import { formatPercent } from "metabase/static-viz/lib/numbers";
+import { formatPercent } from "metabase/utils/formatting";
 import type { TooltipRowModel } from "metabase/visualizations/types";
 
 import S from "./TooltipRow.module.css";
