+++ frontend/src/metabase/visualizations/echarts/graph/treemap/option/option.unit.spec.ts
@@ -1,4 +1,4 @@
-import { formatPercent } from "metabase/static-viz/lib/numbers";
+import { formatPercent } from "metabase/utils/formatting";
 import { DEFAULT_VISUALIZATION_THEME } from "metabase/visualizations/shared/utils/theme";
 import type { RenderingContext } from "metabase/visualizations/types";
